+CmdletStepContext to help invoke CmdletHost
There's an inherent conflict from the CmdletHost API and reusing steps. For CmdletHost, you commit up front to what kind of Cmdlet you're hosting--and then you use that type throughout the invocation. To reuse steps though, we need to be fluid with what kind of Cmdlet we're hosting, since it can change between scenarios. There's an impedance here.

This class tries to bridge the gap (without reflection!), by allowing the steps to not ENTIRELY commit up front to the Cmdlet they're using. Instead, this class adapts the host API and adds wiggle room for changing what Cmdlet type it is.