Pushing the "return value" idea into the context
This enables the steps to be completely focused on the context, and doesn't have to maintain its own state. Instead, the context handles all the state.

Also re-thought the concept, instead of return value, it's more like the cmdlet output, since it's likely writing something to the object pipeline.

Was also able to ditch an unused step that was too specialized.